got good v & s values. put turn pid on motor.
